Fixed a bunch of wvstreams tests for 64-bit Linux.  (sizeof(long) == sizeof(void *) under 64-bit; sizeof(int) != sizeof(void *), so it would be hard to case void *'s to ints).

+++ utils/t/wvcallback.t.cc
@@ -5,8 +5,8 @@
 // START callbacktest.cc definitions
 struct A
 {
-    int x, y;
-    A(int _x = 0, int _y = 0)
+    long x, y;
+    A(long _x = 0, long _y = 0)
         { x = _x; y = _y; }
    
     A add(const A &a)
@@ -20,7 +20,7 @@ typedef wv::function<A(void*)> A3Callback;
 
 static A bunk(const A &a, void *userdata)
 {
-    int incr = (int)userdata;
+    long incr = (long)userdata;
     return A(a.x+incr, a.y+incr*2);
 }
