+++ utils/tests/callbacktest.cc
@@ -3,8 +3,8 @@
 
 struct A
 {
-    int x, y;
-    A(int _x = 0, int _y = 0)
+    long x, y;
+    A(long _x = 0, long _y = 0)
         { x = _x; y = _y; }
     
     A add(const A &a)
@@ -18,7 +18,7 @@ typedef wv::function<A(void *)> A3Callback;
 
 static A bunk(const A &a, void *userdata)
 {
-    int incr = (int)userdata;
+    long incr = (long)userdata;
     return A(a.x+incr, a.y+incr*2);
 }
